Add views for pcpapillon api
